requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.3
